+++ acs.php
@@ -28,7 +28,9 @@ global $db, $conf, $langs, $hookmanager;
 require_once DOL_DOCUMENT_ROOT.'/core/class/html.formfile.class.php';
 require_once __DIR__.'/lib/autoload.php';
 
-$login = get_saml();
+$fk_idp = intval(GETPOST('fk_idp', 'int'));
+
+$login = get_saml($fk_idp);
 
 $login->processResponse();
 
@@ -42,9 +44,10 @@ if($login->isAuthenticated()) {
     if(! empty($conf->global->SAMLCONNECTOR_CREATE_UNEXISTING_USER) || ! empty($conf->global->SAMLCONNECTOR_UPDATE_USER_EVERYTIME)) {
         $user->firstname = $login->getAttribute($conf->global->SAMLCONNECTOR_MAPPING_USER_FIRSTNAME)[0];
         $user->lastname = $login->getAttribute($conf->global->SAMLCONNECTOR_MAPPING_USER_LASTNAME)[0];
-        $user->admin = in_array('ADMINISTRATOR', $login->getAttribute('type')) ? 1 : 0;
+//        $user->admin = in_array('ADMINISTRATOR', $login->getAttribute('type')) ? 1 : 0;
         $user->email = $login->getAttribute($conf->global->SAMLCONNECTOR_MAPPING_USER_EMAIL)[0];
     }
+
 
     if($res <= 0 && ! empty($conf->global->SAMLCONNECTOR_CREATE_UNEXISTING_USER)) {
         $user->login = $login->getNameId();
@@ -66,10 +69,13 @@ if($login->isAuthenticated()) {
         $_SESSION['dol_screenwidth'] = $dol_screenwidth ?? '';
         $_SESSION['dol_screenheight'] = $dol_screenheight ?? '';
         $_SESSION['dol_company'] = $conf->global->MAIN_INFO_SOCIETE_NOM;
-        $_SESSION['dol_entity'] = $conf->entity;
+        $_SESSION['dol_samlconnector_fk_idp'] = $fk_idp;
+        if(GETPOSTISSET('entity')) $_SESSION['dol_entity'] = GETPOST('entity','int');
+        else $_SESSION['dol_entity'] = $conf->entity;
+
         if (isModEnabled('multicompany')) {
             $entityFromDB = null;
-            $sql = "SELECT entity FROM " . MAIN_DB_PREFIX . "user WHERE login = '" . $db->escape($user->login) . "' AND statut = 1";
+            $sql = "SELECT entity FROM " . $db->prefix() . "user WHERE login = '" . $db->escape($user->login) . "' AND statut = 1";
             $resql = $db->query($sql);
             if ($resql) {
                 $obj = $db->fetch_object($resql);
@@ -77,6 +83,9 @@ if($login->isAuthenticated()) {
             }
             if ($entityFromDB) $_SESSION['dol_entity'] = $entityFromDB;
         }
+
+
+
         // Store value into session (values stored only if defined)
         if(! empty($dol_hide_topmenu)) $_SESSION['dol_hide_topmenu'] = $dol_hide_topmenu;
         if(! empty($dol_hide_leftmenu)) $_SESSION['dol_hide_leftmenu'] = $dol_hide_leftmenu;
@@ -103,13 +112,13 @@ if($login->isAuthenticated()) {
         $reshook = $hookmanager->executeHooks('afterLogin', $parameters, $user, $action);    // Note that $action and $object may have been modified by some hooks
 
         $db->commit();
-        header('Location: '.DOL_URL_ROOT);
     }
 }
 
 if(isset($_REQUEST['RelayState'])) {
     $login->redirectTo($_REQUEST['RelayState']);
+} else {
+    header('Location: '.DOL_URL_ROOT);
+    exit;
 }
 ?>
-<a href="login.php">Login</a>
-<a href="logout.php">Logout</a>