+++ class/actions_samlconnector.class.php
@@ -220,42 +220,86 @@ class ActionsSamlConnector {
      * @param array $parameters
      * @throws \OneLogin\Saml2\Error
      */
-    public function getLoginPageOptions($parameters) {
-        global $langs;
-
-        $url = dol_buildpath('/samlconnector', 1).'/login.php';
-        $this->results = ['table' => '<p><a href="'.$url.'">'.$langs->trans('SamlConnectorConnectBySaml').'</a></p>'];
-
-        //Force login
-        include dirname(__FILE__).'/../lib/autoload.php';
-        $login = get_saml();
-
-        $newpath = DOL_MAIN_URL_ROOT.'/index.php?mainmenu=home&leftmenu=home';
-        $login->login($newpath);
-    }
+	public function getLoginPageOptions($parameters) {
+		global $langs, $conf;
+		$langs->load('samlconnector@samlconnector');
+		if (!empty($conf->global->SAMLCONNECTOR_IDP_DISPLAY_BUTTON)) {
+			if (!empty($conf->global->SAMLCONNECTOR_MANAGE_MULTIPLE_IDP)) {
+				dol_include_once('samlconnector/class/samlconnectoridp.class.php');
+				$idp = new SamlConnectorIDP($this->db);
+				$idp->ismultientitymanaged = 0;
+				$TIdps = $idp->fetchAll();
+				if (is_array($TIdps) && !empty($TIdps)) {
+					$this->resprints = '<div class="samlConnectorLoginButtonBlock">';
+					foreach ($TIdps as $idp) {
+						$moreclass = '';
+						if ($idp->status == SamlConnectorIDP::STATUS_INACTIVE) continue;
+						if (!empty($conf->multicompany->enabled) && $conf->entity != $idp->entity) $moreclass .= 'hidden';
+						$this->resprints .= '<div class="samlConnectorLoginButtonElement ' . $moreclass . '" data-entity="' . $idp->entity . '">';
+						$this->resprints .= $idp->getLoginButton();
+						$this->resprints .= '</div>';
+					}
+					$this->resprints .= '</div>
+				<script type="text/javascript">
+					$(".samlConnectorLoginButtonBlock").appendTo("#login-submit-wrapper");
+                    $("#entity").on("change", function (){
+                        let entity = $(this).val();
+                        $(\'div[data-entity="\'+entity+\'"]\').removeClass("hidden");
+                        $(\'div[data-entity][data-entity!="\'+entity+\'"]\').addClass("hidden");
+                    })
+				</script>';
+				}
+			} else {
+				$url = dol_buildpath('/samlconnector', 1).'/login.php';
+				$this->resprints = '<p><a href="'.$url.'">'.$langs->trans('SamlConnectorConnectBySaml').'</a></p>';
+			}
+		}
+		else {
+			//Force login
+			include dirname(__FILE__).'/../lib/autoload.php';
+			try {
+				$login = get_saml() ;
+			}
+			catch(Exception $e) {
+				setEventMessage($langs->trans('PleaseConfigureSamlConnectorModule', $e->getMessage()), 'errors');
+			}
+			finally {
+				$newpath = $_SERVER['REQUEST_URI'];
+				if(!empty($login)) $login->login($newpath);
+			}
+		}
+	}
 
     /**
      * @param array $parameters
      * @throws \OneLogin\Saml2\Error
-     */
-    public function afterLogout($parameters) {
-        global $conf;
+	 */
+	public function afterLogout($parameters)
+	{
+		global $conf;
 
-        if($_SESSION['dol_authmode'] != 'saml') return;
+		if ($_SESSION['dol_authmode'] != 'saml') return;
 
-        $urlfrom = empty($_SESSION['urlfrom']) ? '' : $_SESSION['urlfrom'];
+		$urlfrom = empty($_SESSION['urlfrom']) ? '' : $_SESSION['urlfrom'];
 
-        $url = DOL_URL_ROOT.'/index.php';        // By default, go to login page
-        if($urlfrom) $url = DOL_URL_ROOT.$urlfrom;
-        if(! empty($conf->global->MAIN_LOGOUT_GOTO_URL)) $url = $conf->global->MAIN_LOGOUT_GOTO_URL;
+		$url = DOL_URL_ROOT . '/index.php';        // By default, go to login page
+		if ($urlfrom) $url = DOL_URL_ROOT . $urlfrom;
+		if (!empty($conf->global->MAIN_LOGOUT_GOTO_URL)) $url = $conf->global->MAIN_LOGOUT_GOTO_URL;
+		$fk_idp = intval($_SESSION['dol_samlconnector_fk_idp']);
 
-        foreach(array_keys($_SESSION) as $key) {
-            unset($_SESSION[$key]);
-        }
+		foreach (array_keys($_SESSION) as $key) {
+			unset($_SESSION[$key]);
+		}
+		if (empty($conf->global->SAMLCONNECTOR_DISABLE_IDP_DISCONNECTION)) {
+			include dirname(__FILE__) . '/../lib/autoload.php';
+			$login = get_saml($fk_idp);
+			$login->logout($url);
+		}
+	}
 
-        include dirname(__FILE__).'/../lib/autoload.php';
-
-        $login = get_saml();
-        $login->logout($url);
-    }
+	public function selectForFormsListWhere($parameters, &$object, &$action, $hookmanager) {
+		if($parameters['currentcontext'] == 'samlconnectorsetup') {
+			$this->resprints = ' WHERE t.active = 1';
+		}
+	}
 }