+++ css/samlconnector.css.php
@@ -22,14 +22,14 @@ if(! defined('NOLOGIN')) define('NOLOGIN', 1);
 if(! defined('NOREQUIREHTML')) define('NOREQUIREHTML', 1);
 if(! defined('NOREQUIREAJAX')) define('NOREQUIREAJAX', 1);
 
+session_cache_limiter(false);
+
 // Load Dolibarr environment
 $res = 0;
 $main_inc = 'main.inc.php';
 for($i = 0 ; $i < 5 && ! $res ; $i++) $res = @include str_repeat('../', $i).$main_inc;
 
 require_once DOL_DOCUMENT_ROOT.'/core/lib/functions2.lib.php';
-
-session_cache_limiter(false);
 // Define css type
 header('Content-type: text/css');
 // Important: Following code is to cache this file to avoid page request by browser at each Dolibarr page access.
@@ -58,4 +58,40 @@ else header('Cache-Control: no-cache');
     vertical-align: middle;
     max-height: 16px;
     max-width: 16px;
+}
+
+.samlConnectorLoginButton {
+    display: inline-block;
+    width: 100%;
+    height: 100%;
+    text-decoration: none !important;
+    color: black !important;
+
+}
+
+.samlConnectorLoginButtonBlock {
+    text-align: center;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+}
+
+.samlConnectorLoginButtonElement {
+    min-width: min(300px, 100%);
+    /*border: 1px solid transparent;*/
+    padding: 8px 10px;
+
+    background-color: white;
+    color: black;
+    margin-top: 15px;
+    font-size: 16px;
+    cursor: pointer;
+    box-shadow: rgb(0 0 0 / 20%) 1px 1px 5px 0;
+    transition: background-color .2s ease-in-out 0s,border-color .2s ease-in-out 0s;
+}
+
+.samlConnectorLoginButton img{
+    float: left;
+    max-height: 21px;
 }